+++ src/Sentry/SentryOptions.cs
@@ -853,7 +853,7 @@ public class SentryOptions
     public TimeSpan InitCacheFlushTimeout { get; set; } = TimeSpan.FromSeconds(1);
 
     /// <summary>
-    /// Defaults tags to add to all events. (These are indexed by Sentry).
+    /// Default tags to add to all events. (These are indexed by Sentry).
     /// </summary>
     /// <remarks>
     /// If the key already exists in the event, it will not be overwritten by a default tag.
@@ -980,8 +980,7 @@ public class SentryOptions
     /// <summary>
     /// A customizable list of <see cref="StringOrRegex"/> objects, each containing either a
     /// substring or regular expression pattern that can be used to control which outgoing HTTP requests
-    /// will have the <c>sentry-trace</c>, <c>traceparent</c>, and <c>baggage</c> headers propagated,
-    /// for purposes of distributed tracing.
+    /// will have the <c>sentry-trace</c> and <c>baggage</c> headers propagated, for purposes of distributed tracing.
     /// The default value contains a single value of <c>.*</c>, which matches everything.
     /// To disable propagation completely, clear this collection or set it to an empty collection.
     /// </summary>
@@ -1321,48 +1320,7 @@ public class SentryOptions
         Native = new NativeOptions(this);
 #endif
 
-        InAppExclude = new() {
-                "System",
-                "Mono",
-                "Sentry",
-                "Microsoft",
-                "MS", // MS.Win32, MS.Internal, etc: Desktop apps
-                "ABI.Microsoft", // MAUI
-                "WinRT", // WinRT, UWP, WinUI
-                "UIKit", // iOS / MacCatalyst
-                "Newtonsoft.Json",
-                "FSharp",
-                "Serilog",
-                "Giraffe",
-                "NLog",
-                "Npgsql",
-                "RabbitMQ",
-                "Hangfire",
-                "IdentityServer4",
-                "AWSSDK",
-                "Polly",
-                "Swashbuckle",
-                "FluentValidation",
-                "Autofac",
-                "Stackexchange.Redis",
-                "Dapper",
-                "RestSharp",
-                "SkiaSharp",
-                "IdentityModel",
-                "SqlitePclRaw",
-                "Xamarin",
-                "Android", // Ex: Android.Runtime.JNINativeWrapper...
-                "Google",
-                "MongoDB",
-                "Remotion.Linq",
-                "AutoMapper",
-                "Nest",
-                "Owin",
-                "MediatR",
-                "ICSharpCode",
-                "Grpc",
-                "ServiceStack"
-        };
+        InAppExclude = GetDefaultInAppExclude();
 
 #if DEBUG
         InAppInclude = new()
@@ -1856,4 +1814,49 @@ public class SentryOptions
         // In the future, this will most likely contain process ID
         return TryGetDsnSpecificCacheDirectoryPath();
     }
+
+    internal static List<StringOrRegex> GetDefaultInAppExclude() =>
+    [
+        "System",
+        "Mono",
+        "Sentry",
+        "Microsoft",
+        "MS", // MS.Win32, MS.Internal, etc: Desktop apps
+        "ABI.Microsoft", // MAUI
+        "WinRT", // WinRT, UWP, WinUI
+        "UIKit", // iOS / MacCatalyst
+        "Newtonsoft.Json",
+        "FSharp",
+        "Serilog",
+        "Giraffe",
+        "NLog",
+        "Npgsql",
+        "RabbitMQ",
+        "Hangfire",
+        "IdentityServer4",
+        "AWSSDK",
+        "Polly",
+        "Swashbuckle",
+        "FluentValidation",
+        "Autofac",
+        "Stackexchange.Redis",
+        "Dapper",
+        "RestSharp",
+        "SkiaSharp",
+        "IdentityModel",
+        "SqlitePclRaw",
+        "Xamarin",
+        "Android", // Ex: Android.Runtime.JNINativeWrapper...
+        "Google",
+        "MongoDB",
+        "Remotion.Linq",
+        "AutoMapper",
+        "Nest",
+        "Owin",
+        "MediatR",
+        "ICSharpCode",
+        "Grpc",
+        "ServiceStack",
+        "Java.Interop",
+    ];
 }