+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 5.8.2-beta.1
+## Unreleased
 
 ### Fixes